Revert "Add Contact, Orders page and routing"
