+++ src/App.js
@@ -1,20 +1,13 @@
-import React, { Component } from "react";
-import Layout from "./hoc/Layout/Layout";
-import BurgerBuilder from "./containers/BurgerBuilder/BurgerBuilder";
-import Checkout from "./containers/Checkout/Checkout";
-import { Route, Switch } from "react-router-dom";
-import Orders from './containers/Orders/Orders';
+import React, { Component } from 'react';
+import Layout from './hoc/Layout/Layout';
+import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder';
 
 class App extends Component {
   render() {
     return (
       <div>
         <Layout>
-          <Switch>
-            <Route path="/checkout" component={Checkout} />
-            <Route path='/orders' component={Orders} />
-            <Route path="/" exact component={BurgerBuilder} />
-          </Switch>
+          <BurgerBuilder />
         </Layout>
       </div>
     );
