+++ src/components/Burger/BuildControls/BuildControl/BuildControl.module.css
@@ -29,7 +29,7 @@
   cursor: not-allowed;
 }
 
-.BuildControl .Label {
+.Label {
   padding: 10px;
   font-weight: bold;
   width: 80px;
