+++ src/components/Burger/OrderSummary/OrderSummary.js
@@ -1,4 +1,5 @@
 import React, {useEffect} from "react";
+import Auxiliary from "../../../hoc/Auxiliary/Auxiliary";
 import Button from '../../UI/Button/Button';
 
 const OrderSummary = (props) => {
@@ -20,7 +21,7 @@ const OrderSummary = (props) => {
   });
 
   return (
-    <React.Fragment>
+    <Auxiliary>
       <h3>Your Order</h3>
       <p>A delicious burger with the following ingredients:</p>
       <ul>
@@ -30,7 +31,7 @@ const OrderSummary = (props) => {
       <p>Continue to checkout?</p>
       <Button onClick={props.onModalClosed} btnType="Danger">CANCEL</Button>
       <Button onClick={props.onContinuePurchasing} btnType="Success">CONTINUE</Button>
-    </React.Fragment>
+    </Auxiliary>
   );
 };
