+++ src/components/Navigation/NavigationItems/NavigationItem/NavigationItem.js
@@ -1,10 +1,11 @@
 import React from "react";
 import classes from "./NavigationItem.module.css";
-import { NavLink } from "react-router-dom";
 
 const navigationItem = (props) => (
   <li className={classes.NavigationItem}>
-    <NavLink exact activeClassName={classes.active} to={props.link}>{props.children}</NavLink>
+    <a className={props.active ? classes.active : null} href={props.link}>
+      {props.children}
+    </a>
   </li>
 );
