+++ src/components/Navigation/NavigationItems/NavigationItems.js
@@ -4,8 +4,8 @@ import NavigationItem from "./NavigationItem/NavigationItem";
 
 const navigationItems = () => (
   <ul className={classes.NavigationItems}>
-    <NavigationItem link="/">Burger Builder</NavigationItem>
-    <NavigationItem link="/orders">Orders</NavigationItem>
+    <NavigationItem link="/" active>Burger Builder</NavigationItem>
+    <NavigationItem link="/">Checkout</NavigationItem>
   </ul>
 );
