+++ src/components/Navigation/SideDrawer/SideDrawer.js
@@ -3,6 +3,7 @@ import Logo from "../../Logo/Logo";
 import NavigationItems from "../NavigationItems/NavigationItems";
 import classes from "./SideDrawer.module.css";
 import Backdrop from "../../UI/Backdrop/Backdrop";
+import Auxiliary from "../../../hoc/Auxiliary/Auxiliary";
 
 const sideDrawer = (props) => {
 
@@ -12,7 +13,7 @@ const sideDrawer = (props) => {
   }
 
   return (
-    <React.Fragment>
+    <Auxiliary>
       <Backdrop clicked={props.onSideDrawerClose} show={props.open} />
       <div className={attachedClasses.join(' ')}>
         <div className={classes.Logo}>
@@ -22,7 +23,7 @@ const sideDrawer = (props) => {
           <NavigationItems />
         </nav>
       </div>
-    </React.Fragment>
+    </Auxiliary>
   );
 };
