--- src/components/Order/CheckoutSummary/CheckoutSummary.module.css
@@ -1,16 +1,0 @@
-.CheckoutSummary {
-  text-align: center;
-  width: 80%;
-  margin: auto;
-}
-
-.CheckoutSummary .Burger {
-  width: 100%;
-  margin: auto;
-}
-
-@media (min-width: 600px) {
-  .CheckoutSummary {
-    width: 500px;
-  }
-}
