--- src/components/Order/Order.module.css
@@ -1,8 +1,0 @@
-.Order {
-  width: 80%;
-  border: 1px solid #eee;
-  box-shadow: 0 2px 3px #ccc;
-  padding: 10px;
-  margin: 10px auto;
-  box-sizing: border-box;
-}
