+++ src/containers/BurgerBuilder/BurgerBuilder.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import Auxiliary from "../../hoc/Auxiliary/Auxiliary";
 import Burger from "../../components/Burger/Burger";
 import BuildControls from "../../components/Burger/BuildControls/BuildControls";
 import Modal from "../../components/UI/Modal/Modal";
@@ -15,13 +16,16 @@ const INGREDIENT_PRICES = {
 };
 
 class BurgerBuilder extends Component {
-  state = {
-    ingredients: null,
-    totalPrice: 4,
-    purchasable: false,
-    purchasing: false,
-    loading: false,
-  };
+  constructor(props) {
+    super(props);
+    this.state = {
+      ingredients: null,
+      totalPrice: 4,
+      purchasable: false,
+      purchasing: false,
+      loading: false,
+    };
+  }
 
   handleModalClose = () => {
     this.setState({ purchasing: false });
@@ -90,30 +94,35 @@ class BurgerBuilder extends Component {
   };
 
   handleContinuePurchasing = () => {
-    const queryParams = [];
-    for (let i in this.state.ingredients) {
-      queryParams.push(
-        encodeURIComponent(i) +
-          "=" +
-          encodeURIComponent(this.state.ingredients[i])
-      );
-    }
-    queryParams.push('price=' + this.state.totalPrice);
-    const queryString = queryParams.join("&");
-    this.props.history.push({
-      pathname: '/checkout',
-      search: "?" + queryString,
-    });
+    this.setState({ loading: true });
+    const order = {
+      ingredients: this.state.ingredients,
+      price: this.state.totalPrice,
+      customer: {
+        name: "Starea",
+        address: {
+          street: "NY",
+          zipCode: "123456",
+          country: "Vietnam",
+        },
+        email: "thinhle2199@gmail.com",
+      },
+    };
+    axios
+      .post("/orders.json", order)
+      .then((res) => {
+        this.setState({ loading: false, purchasing: false });
+      })
+      .catch((err) => {
+        this.setState({ loading: false, purchasing: false });
+      });
   };
 
   componentDidMount() {
-    console.log("[BurgerBuilder.js] did mount");
-    axios
-      .get("/ingredients.json")
-      .then((res) => {
-        this.setState({ ingredients: res.data });
-      })
-      .catch((err) => {});
+    console.log('[BurgerBuilder.js] did mount');
+    axios.get("/ingredients.json").then((res) => {
+      this.setState({ ingredients: res.data });
+    }).catch(err => {});
   }
 
   render() {
@@ -133,7 +142,7 @@ class BurgerBuilder extends Component {
 
     let burger = <Spinner />;
 
-    if (this.state.ingredients) {
+    if(this.state.ingredients) {
       orderSummary = (
         <OrderSummary
           totalPrice={this.state.totalPrice}
@@ -144,7 +153,7 @@ class BurgerBuilder extends Component {
       );
 
       burger = (
-        <React.Fragment>
+        <Auxiliary>
           <Burger ingredients={this.state.ingredients} />
           <BuildControls
             price={this.state.totalPrice}
@@ -154,7 +163,7 @@ class BurgerBuilder extends Component {
             purchasable={this.state.purchasable}
             onOrderButtonClick={this.handleOrderButtonClick}
           />
-        </React.Fragment>
+        </Auxiliary>
       );
     }
 
@@ -163,7 +172,7 @@ class BurgerBuilder extends Component {
     }
 
     return (
-      <React.Fragment>
+      <Auxiliary>
         <Modal
           onModalClosed={this.handleModalClose}
           show={this.state.purchasing}
@@ -171,7 +180,7 @@ class BurgerBuilder extends Component {
           {orderSummary}
         </Modal>
         {burger}
-      </React.Fragment>
+      </Auxiliary>
     );
   }
 }
