--- src/containers/Checkout/ContactData/ContactData.js
@@ -1,89 +1,0 @@
-import React, { Component } from "react";
-import Button from "../../../components/UI/Button/Button";
-import classes from "./ContactData.module.css";
-import axios from "../../../axios-orders";
-import Spinner from "../../../components/UI/Spinner/Spinner";
-
-class ContactData extends Component {
-  state = {
-    name: "",
-    email: "",
-    address: {
-      street: "",
-      postalCode: "",
-    },
-    loading: false,
-  };
-
-  handleOrder = (e) => {
-    e.preventDefault();
-    this.setState({ loading: true });
-    const order = {
-      ingredients: this.props.ingredients,
-      price: this.props.price,
-      customer: {
-        name: this.state.name,
-        address: {
-          street: "NY",
-          zipCode: "123456",
-          country: "Vietnam",
-        },
-        email: "thinhle2199@gmail.com",
-      },
-    };
-    axios
-      .post("/orders.json", order)
-      .then((res) => {
-        this.setState({ loading: false });
-        this.props.history.push('/');
-      })
-      .catch((err) => {
-        this.setState({ loading: false });
-      });
-  };
-
-  render() {
-    let form = (
-      <form>
-        <input
-          className={classes.Input}
-          type="text"
-          name="name"
-          placeholder="Your name"
-        />
-        <input
-          className={classes.Input}
-          type="text"
-          name="email"
-          placeholder="Your email"
-        />
-        <input
-          className={classes.Input}
-          type="text"
-          name="street"
-          placeholder="Street"
-        />
-        <input
-          className={classes.Input}
-          type="text"
-          name="postalCode"
-          placeholder="Postal code"
-        />
-        <Button onClick={this.handleOrder} btnType="Success">
-          ORDER
-        </Button>
-      </form>
-    );
-    if (this.state.loading) {
-      form = <Spinner />;
-    }
-    return (
-      <div className={classes.ContactData}>
-        <h4>Enter your Contact Data</h4>
-        {form}
-      </div>
-    );
-  }
-}
-
-export default ContactData;
