--- src/containers/Checkout/ContactData/ContactData.module.css
@@ -1,18 +1,0 @@
-.ContactData {
-  margin: 20px auto;
-  width: 80%;
-  text-align: center;
-  box-shadow: 0 2px 3px #ccc;
-  padding: 10px;
-  box-sizing: border-box;
-}
-
-.ContactData .Input {
-  display: block;
-}
-
-@media (min-width: 600px) {
-  .ContactData {
-    width: 500px;
-  }
-}
