+++ src/hoc/Auxiliary/Auxiliary.js
@@ -1,0 +1,3 @@
+const auxiliary = (props) => props.children;
+
+export default auxiliary;
