+++ src/hoc/Layout/Layout.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import Auxiliary from "../Auxiliary/Auxiliary";
 import classes from "./Layout.module.css";
 import Toolbar from "../../components/Navigation/Toolbar/Toolbar";
 import SideDrawer from "../../components/Navigation/SideDrawer/SideDrawer";
@@ -19,11 +20,11 @@ class Layout extends Component {
 
   render() {
     return (
-      <React.Fragment>
+      <Auxiliary>
         <Toolbar onSideDrawerOpen={this.handleSideDrawerOpen} />
         <SideDrawer open={this.state.showSideDrawer} onSideDrawerClose={this.handleSideDrawerClose} />
         <main className={classes.Content}>{this.props.children}</main>
-      </React.Fragment>
+      </Auxiliary>
     );
   }
 }
