+++ src/hoc/withErrorHandler/withErrorHandler.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import Modal from "../../components/UI/Modal/Modal";
+import Auxiliary from "../Auxiliary/Auxiliary";
 
 const withErrorHandler = (WrappedComponent, axios) => {
   return class extends Component {
@@ -32,7 +33,7 @@ const withErrorHandler = (WrappedComponent, axios) => {
 
     render() {
       return (
-        <React.Fragment>
+        <Auxiliary>
           <Modal
             onModalClosed={this.errorConfirmHandler}
             show={this.state.error}
@@ -40,7 +41,7 @@ const withErrorHandler = (WrappedComponent, axios) => {
             {this.state.error ? this.state.error.message : null}
           </Modal>
           <WrappedComponent {...this.props} />
-        </React.Fragment>
+        </Auxiliary>
       );
     }
   };
